requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
streamlit
